Version 2.0
Give same behavior to all paths, this means that links, grep results paths, recursive listing paths all behave the same way like files in the directory listing. So now it is possible for example to open a link in the favorites bar or history bar in another tab using <space>t or to open a grep result in a vertical split window doing <space>v. See the "File operations" section in the documentation usage, to see other possible actions to do on directories, files and links.
Now directory may be opened in new tabs or new split etc like any file.
The configuration file is now a xml file the previous file format using a dictionary is not compatible anymore. The plugin used for this xml file is "basicXmlParser.vim" which is included in this plugin so that each release will have the corresponding working version of the plugin
Copy and move files and directories works in windows (not tested in linux)
Now possible to open multiple selected files each in a new tab
Added shift-j To open the next file in the preview window, shift-k Go open the previous file in the preview window, and also the default action now to open items is in the preview window
It is now possible to add files to the favorites bar using <space>B (<space>b add a directory). This means that the favorites bar can now be used to bookmark directories but also to contain links to files (like shortcuts).
Corrected bug in the GetFileName() function when there was a filter set with <space>f, the files could not opened.
I removed one function that was not used anymore to add favorites, it was replaced by AddToBar() which is used also for the history bar.
Removed the VeGetFileName() function, put its code in the VeGetPath() function
Corrected a bug in the change directory function, now changes directory without interruption
Change <space>s (save configuration) for <space>S. <space>s is not used to select a file.
Added <space>X Open current directory with the gvim "browse" command to open a file in the current buffer
Changed the fold level for the mappings so they appear at the first level for more easy access
Put the matchadd() to set the colors inside try..catch in case the highlight groups defined are not present in the current color scheme. I added a documentation section explaining how to change these colors.
Added a new variable to know what was the last history path added not to add the current path multiple times (suggestion from Xaizek)
Replaced the fixed value for the starting line of the directory listing by a function called "g:VeGetDirectoryListLineNum()"
Removed a "\s" from the regexes to test if the line is a directory or a file ("^d.\{9}\s" or "^-.\{9}\s") because in cygwin there is only one \s but with UnxUtils there are 2.
Changed the variable $home by $HOME because in linux it is case sensitive
Removed the option "setlocal nowrap" in the BuildWindow() function, because a user may want always to have the history and favorites bar wrapped
Corrected the <leader>VV mapping, it was <leader>VV> and was not executing
Corrected/added <space>` to show the marks same like <space>'
Added user-commands that can be run on selected files
I remove TabEnter/Leave and BufEnter/Leave from the autocommands which call s:OnEnter() and s:OnLeave(), this was causing these functions to be called 3 times thus slowing access to the windows, this removed the necessity for the usage of the variable g:VeAutoRefreshOnEnter that was removed
Change map to nmap in mappings to start the plugin
Added mappings to inspect the content of the selected files object, the commands object and the command to run object
Added operations on selected files in a recursive listing also (ls -R) Copy, move, delete, OpenInNewTab etc
Added selection of items in the links bar