Version 2.1
Corrected bug in the s:SetCursorPos(path) (to exit if there are no path)
Did testing when there are no VimExplorer.xml file
Change suggested mappings (commented out) (f2, f3, ..., Home, PageUp, etc)
Modified the installation instructions to indicated that if you use the plugin in Windows, you will have to change the value of the "g:VeLsCmdPath" to that of the path where your ls command is located. Search for the variable using: /let g:VeLsCmdPath
Added <space>9 If Windows is used to change the ls command for "UnxUtils" ls command
Added <space>0 If Windows is used to Change the ls command for "Cygwin" ls command
Modified also the installation instructions to indicated that if you use the plugin in Windows, you will have to change also the value of the "g:VeGrepCmdPath" to that of the path where your grep command is located. Search for the variable using: /let g:VeGrepCmdPath (You may also set that variable to 'grep' simply and set the path to the executable using the system %path% variable)
Change the regex so that the cursor position is remembered also when there is a filter: changed "cal search('\([0-9]\s\+\|\/\)\zs\s\+' . g:VeCfg.CursorPos[g:VePath].Value . '$')" to "cal search('[0-9]\zs\s\+' . g:VeCfg.CursorPos[g:VePath].Value . '$')"